council_orchestrator/requirements.txt

groq
python-dotenv
